+++ app/api/src/main/java/org/cart/next/Repositories/PedidoRepository.java
@@ -1,5 +1,6 @@
 package org.cart.next.Repositories;
 
+import java.math.BigInteger;
 import java.util.List;
 import org.cart.next.Entities.Pedido;
 
@@ -7,7 +8,7 @@ public interface PedidoRepository {
 
     public List<Pedido> obterPedidos();
     
-    public Pedido obterPedidoPeloId(Long id);
+    public Pedido obterPedidoPeloId(BigInteger id);
 
     public void salvarPedido(Pedido pedido);
 }
