+++ app/api/src/main/java/org/cart/next/Repositories/ProdutoRepository.java
@@ -1,5 +1,6 @@
 package org.cart.next.Repositories;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.cart.next.Entities.Produto;
@@ -8,5 +9,5 @@ public interface ProdutoRepository {
     
     public List<Produto> obterProdutos();
 
-    public Produto obterProdutoPeloId(Long id);
+    public Produto obterProdutoPeloId(BigInteger id);
 }
